1.3: Remove unicode, more documentation, clearer examples.
- Remove unicode symbols, documentation, code clarifications.

Basically I did a lot more explicit building of `Lens` family types to
show what each component does. Then I wrote a lot of
documentation. And also pulled out the conditional from `main` (the
`case` statement) to keep as much as possible inside `Applicative`.

- Full Haddock Coverage now, even generates source.

- Print a message on start-up about how to view the application in a browser.

The user can now click on the link and see the program in the browser,
live.

- Explain more about MonadIO, Lawless.